Change default color theme trace values and add Graph Styling docs
